Fix critical bugs in contracts and frontend
Contracts:
- PoolManager: fix swap() slippage direction (was adding instead of
  subtracting), apply README midpoint formula p1*2s/(2s+ratio)
- PoolManager: fix EASY reward decimal calculation in swap()
- Pool: fix uint32 truncation for 40-bit settings entries (use uint64)
- Governance: prevent double-voting with per-proposal voted mapping
- SimplePriceHelper: add Ownable, restrict setMockPrice to onlyOwner

Frontend:
- contracts.js: rewrite PoolManager ABI (was missing addLiquidity/
  removeLiquidity/swap/getELFPrice, had many non-existent functions)
- contracts.js: rewrite Governance ABI (was wrong OpenZeppelin Governor
  interface instead of actual contract interface)
- contracts.js: fix ELFToken ABI (remove fake pause/unpause, add real fns)
- app.js: executeSwap now calls approve + poolManager.swap + pool.withdraw
- app.js: addLiquidity now calls approve + poolManager.addLiquidity
- app.js: updateSwapInterface reads on-chain TVL instead of Math.random()

Deployment:
- deploy-simple.js: rewrite with correct flow (setPoolManager on ELFToken,
  addPool for each token, seed liquidity via poolManager, auto-update
  frontend/src/contracts.js addresses after deploy)
- hardhat.config.js: add explicit network configs, remove unused plugin
- package.json: add deploy:hardhat and frontend npm scripts
- .gitignore: add node_modules, deployments, frontend/frontend.log

+++ contracts/Pool.sol
@@ -94,7 +94,8 @@ contract Pool is Ownable, IPool {
         // Validate slippage targets are in reasonable range (0-100%)
         uint128 slippageTargets = uint128((newSettings & SETTINGS_TARGET_SLIPPAGES) >> 40);
         for(uint8 i = 0; i < 3; i++){
-            uint32 singleTargetSetting = uint32(slippageTargets & TARGET_SLIPPAGES_SINGLE_MASK);
+            // Each entry is 40 bits (5 bytes): lower 8 bits = ratio, upper 32 bits = edge
+            uint64 singleTargetSetting = uint64(slippageTargets & TARGET_SLIPPAGES_SINGLE_MASK);
             uint8 ratio = uint8(singleTargetSetting & 0xFF);
             require(ratio <= 100, "Slippage ratio cannot exceed 100%");
             slippageTargets = slippageTargets >> SINGLE_TARGET_SETTINGS_LENGTH;
@@ -182,7 +183,7 @@ contract Pool is Ownable, IPool {
         uint8 T = 0; // target Ratio, decimal of T is 2
         
         for(uint8 i = 0; i < 3; i++){
-            uint32 singleTargetSetting = uint32(slippageTargets & TARGET_SLIPPAGES_SINGLE_MASK);
+            uint64 singleTargetSetting = uint64(slippageTargets & TARGET_SLIPPAGES_SINGLE_MASK);
             (high, T) = getSingleTargetSetting(singleTargetSetting);
             
             if(poolValue >= low && poolValue < high){
@@ -214,10 +215,10 @@ contract Pool is Ownable, IPool {
     }
 
     function getSingleTargetSetting(
-        uint32 slippageTargets
+        uint64 slippageTargets
     ) private pure returns (uint32 high, uint8 ratio){
         ratio = uint8(slippageTargets & 0xFF);
-        high = (slippageTargets >> 8);
+        high = uint32(slippageTargets >> 8);
     }
 
     function estimateSlippageByPoolMode(
